Rename NumberFactorizator module to NumberFactorizor. Update tests module accordingly
