+++ src/main/java/org/number_factorizor/Main.java
@@ -1,4 +1,4 @@
-package org.number_factorizator;
+package org.number_factorizor;
 
 import static java.lang.System.exit;
